Fix projects metadata export by splitting client card
